inventory

+++ LMS_CMS_Backend/LMS_CMS_BL/DTO/Inventory/InventoryDetailsGetDTO.cs
@@ -13,8 +13,8 @@ namespace LMS_CMS_BL.DTO.Inventory
     {
         public long ID { get; set; }
         public int Quantity { get; set; }
-        public float Price { get; set; }
-        public float TotalPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal TotalPrice { get; set; }
         public string? InvoiceNumber { get; set; }
         public string? Notes { get; set; }
         public string? BarCode { get; set; }
